pyproject.toml
[project]
name = "agrigravity"
version = "1.0.0"
description = "AI-Powered Agricultural Assistant Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "shapely>=2.0.2",
    "pyproj>=3.6.1",
    "jinja2>=3.1.3",
    "httpx>=0.26.0",
    "requests>=2.31.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
